Refactoring: Detached the sparse DocValuesFacets from standard DocValuesFacets
